requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
python-telegram-bot==20.4
requests==2.31.0
aiofiles==23.1.0
pydantic==1.10.12
pytest==7.4.0
httpx==0.24.1
